requirements.txt
# Core
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Document Parsing
PyMuPDF>=1.23.0
python-docx>=1.0.0

# Web search (free, no API key required)
duckduckgo-search>=7.0.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
tenacity>=8.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
respx>=0.20.0
httpx>=0.27.0
